Move code to generate apub urls into lemmy_api_common
