+++ crates/apub/src/objects/community.rs
@@ -2,8 +2,6 @@ use crate::{
   check_apub_id_valid_with_strictness,
   collections::{community_moderators::ApubCommunityModerators, CommunityContext},
   fetch_local_site_data,
-  generate_moderators_url,
-  generate_outbox_url,
   local_instance,
   objects::instance::fetch_instance_actor_for_object,
   protocol::{
@@ -20,7 +18,7 @@ use activitypub_federation::{
 use activitystreams_kinds::actor::GroupType;
 use chrono::NaiveDateTime;
 use itertools::Itertools;
-use lemmy_api_common::LemmyContext;
+use lemmy_api_common::{generate_moderators_url, generate_outbox_url, LemmyContext};
 use lemmy_db_schema::{
   source::{
     actor_language::CommunityLanguage,
