+++ crates/apub/src/protocol/collections/group_followers.rs
@@ -1,6 +1,5 @@
-use crate::generate_followers_url;
 use activitystreams_kinds::collection::CollectionType;
-use lemmy_api_common::LemmyContext;
+use lemmy_api_common::{generate_followers_url, LemmyContext};
 use lemmy_db_schema::source::community::Community;
 use lemmy_db_views_actor::structs::CommunityFollowerView;
 use lemmy_utils::error::LemmyError;
